+++ src/Plugin/AuditLogger/Watchdog.php
@@ -3,7 +3,9 @@
 namespace Drupal\os2web_audit\Plugin\AuditLogger;
 
 use Drupal\Core\Logger\LoggerChannelFactoryInterface;
+use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
 use Drupal\Core\Plugin\PluginBase;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Stores entities in the database.
@@ -14,7 +16,7 @@ use Drupal\Core\Plugin\PluginBase;
  *   description = @Translation("Store entity data in the database.")
  * )
  */
-class Watchdog extends PluginBase implements AuditLoggerInterface {
+class Watchdog extends PluginBase implements AuditLoggerInterface, ContainerFactoryPluginInterface {
 
   public function __construct(
     array $configuration,
@@ -27,6 +29,20 @@ class Watchdog extends PluginBase implements AuditLoggerInterface {
 
   /**
    * {@inheritdoc}
+   *
+   * @phpstan-param array<string, mixed> $configuration
+   */
+  public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
+    return new static(
+      $configuration,
+      $plugin_id,
+      $plugin_definition,
+      $container->get('logger.factory'),
+    );
+  }
+
+  /**
+   * {@inheritdoc}
    */
   public function log(string $type, int $timestamp, string $message, array $metadata = []): void {
     $data = '';
@@ -35,9 +51,9 @@ class Watchdog extends PluginBase implements AuditLoggerInterface {
     });
 
     $this->logger->get('os2web_audit')->info('%type: %line (%data)', [
-      'type' => $type,
-      'line' => $message,
-      'data' => $data,
+      '%type' => $type,
+      '%line' => $message,
+      '%data' => $data,
     ]);
   }