+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Made Watchdog default logger
+- Updated Watchlog logger
+
 ## [0.1.0] - 2024-10-21
 
 - Moved drush services into own yml file (drush 12)
@@ -28,6 +31,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/OS2web/os2web_audit/compare/develop...HEAD
 [0.1.0]: https://github.com/OS2web/os2web_audit/compare/0.0.3...0.1.0
-[0.0.2]: https://github.com/OS2web/os2web_audit/compare/0.0.2...0.0.3
+[0.0.3]: https://github.com/OS2web/os2web_audit/compare/0.0.2...0.0.3
 [0.0.2]: https://github.com/OS2web/os2web_audit/compare/0.0.1...0.0.2
 [0.0.1]: https://github.com/OS2web/os2web_audit/releases/tag/0.0.1